+++ pl_bolts/datasets/__init__.py
@@ -1,0 +1,6 @@
+from pl_bolts.datasets.cifar10 import CIFAR10, TrialCIFAR10
+
+__all__ = [
+    'CIFAR10',
+    'TrialCIFAR10',
+]
